agency combobox refactor
