+++ src/js/components/bulkDownload/archive/AwardDataArchiveForm.jsx
@@ -6,18 +6,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { currentFiscalYear, earliestFiscalYear } from "helpers/fiscalYearHelper";
-import ComboBox from "components/sharedComponents/ComboBox";
 import FilterSelectionTitle from "../FilterSelectionTitle";
 import AwardTypeToggle from "./filters/AwardTypeToggle";
 import AgencyComboBox from "./filters/AgencyComboBox";
-
-const currentFY = currentFiscalYear();
-const fyOptions = [];
-
-for (let year = currentFY; year >= earliestFiscalYear; year--) {
-    fyOptions.push({ value: year, text: `FY ${year}` });
-}
+import FYComboBox from "./filters/FYComboBox";
 
 const propTypes = {
     filters: PropTypes.object,
@@ -37,24 +29,13 @@ const AwardDataArchiveForm = ({
         requestResults();
     };
 
-    const onFYSelect = (e) => {
-        e.preventDefault();
-        const target = e.target;
-        updateFilter('fy', target.value);
-    };
-
     return (
         <>
             <FilterSelectionTitle type="agencyFy" />
             <form className="archive-form" onSubmit={handleSubmit}>
                 <div className="award-data-archive-form">
                     <AgencyComboBox agencies={agencies} updateFilter={updateFilter} />
-                    <ComboBox
-                        optionsArray={fyOptions}
-                        onSelect={onFYSelect}
-                        label={"Fiscal Year (FY)"}
-                        placeholder={"Select a Fiscal Year"}
-                        defaultValue={`FY ${currentFY}`} />
+                    <FYComboBox updateFilter={updateFilter} />
                 </div>
                 <AwardTypeToggle filters={filters} updateFilter={updateFilter} />
             </form>
